+++ lib/res/utils/web_browser_info.dart
@@ -1,0 +1,2 @@
+export 'web_browser_info_stub.dart'
+    if (dart.library.html) 'web_browser_info_web.dart';